+++ packages/server/src/index.ts
@@ -2,4 +2,5 @@ export * from "@niarada/remedy-common";
 export * from "@niarada/remedy-hypermedia";
 export * from "@niarada/remedy-runtime";
 export * from "@niarada/remedy-template";
+export * from "./cli";
 export * from "./server";